Move unit meta model elements to bamm namespace
ref #91
